feat: refactor imports and logic rewiring
